backend/requirements.txt
flask
flask-cors

python-firebase
requests=2.31.0